+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+v2014.6.0
+---------
+* Move generic tests to pyfrc.tests
+* IsEnabled() is now IsEnabled(tm), to match other signatures
+* Addition of PracticeMatchTestController
+
 v2014.5.5
 ---------
 * Add support for displaying 'label' attribute in sim UI